+++ lib/preferences/preferences-keys.h
@@ -25,8 +25,8 @@
   X(KSIM_NAV, "simNav", BOOL)            \
   X(KGPS_TX, "gpsTX", UINT)              \
   X(KGPS_RX, "gpsRX", UINT)              \
-  X(KLAT_DFL, "defLAT", DOUBLE)          \
-  X(KLON_DFL, "defLON", DOUBLE)          \
+  X(KLAT_DFL, "defLAT", FLOAT)          \
+  X(KLON_DFL, "defLON", FLOAT)          \
   X(KDEF_BRIGT, "defBright", UINT)       \
   X(KVMAX_BATT, "VmaxBatt", FLOAT)       \
   X(KVMIN_BATT, "VminBatt", FLOAT)       \
